applied createListing changes to master from Seamus_Test_Changes
