--- src/components/createListing/amenities.css
@@ -1,69 +1,0 @@
-.amenities-checkbox {
-  height: auto;
-}
-
-.amenities-list {
-  display: flex;
-  justify-content: space-between;
-  margin: 20px 0;
-}
-
-.no-amenities-check {
-  float: left !important;
-  margin-top: 5px !important;
-}
-
-.create-listing-amenity-container {
-  margin: 0 10px;
-}
-
-.create-listing-amenities-container {
-  margin-bottom: 50px;
-}
-
-#create-listing-amenity {
-  cursor: pointer;
-  transition: transform 0.2s;
-}
-
-#create-listing-amenity:hover {
-  transform: scale(1.2);
-}
-
-.create-listing-amenity-text {
-  font-size: 16px !important;
-  font-weight: normal !important;
-}
-
-.create-listing-no-amenities {
-  pointer-events: none;
-  font-size: 25px !important;
-  color: #dbdbdb;
-}
-
-.create-listing-no-amenities.active {
-  color: var(--color-green);
-}
-
-@media screen and (max-width: 650px) {
-  #create-listing-amenity {
-    height: 35px;
-  }
-  #create-listing-amenity-disabled {
-    height: 35px;
-  }
-  .create-listing-amenity-text {
-    font-size: 12px !important;
-    margin-top: 0;
-  }
-}
-
-@media screen and (max-width: 530px) {
-  .amenities-list {
-    flex-wrap: wrap;
-    justify-content: center;
-  }
-  .create-listing-amenity-container {
-    margin: 0 20px;
-  }
-}
