--- src/components/createListing/detailsListing.css
@@ -1,59 +1,0 @@
-.details-wrapper {
-  display: inline-block;
-  width: 50%;
-  align-content: center;
-  margin: 20px 0;
-}
-
-.beds {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.baths {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.maxppl {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.input-label-details {
-  display: flex;
-  align-items: center;
-  font-weight: bold;
-  font-size: 120% !important;
-  padding: 5%;
-  color: #4d4842;
-}
-
-.input-box-details {
-  width: 50%;
-  padding: 7px 15px;
-  background-color: white;
-  border-radius: 10px;
-  margin: 8px 0;
-  border: 2px solid rgb(210 210 210 / 47%);
-  color: #4d4842;
-  text-align: center;
-  font-size: 20px;
-}
-
-.input-box-details:hover {
-  background-color: inherit;
-}
-
-.input-box-details:focus {
-  border: 2px solid var(--color-green);
-}
-
-.create-listing-details-label {
-  font-size: 120% !important;
-  color: #4d4842;
-  font-weight: bold;
-}
