--- src/components/createListing/endingCL.component.js
@@ -1,13 +1,0 @@
-import React, { Component } from "react";
-
-export default class Ending extends Component {
-  render() {
-    return (
-      <div className="overall-ending">
-        <div className="main-text">
-          <div>Thanks for making a listing!</div>
-        </div>
-      </div>
-    );
-  }
-}
