+++ src/index.js
@@ -4,6 +4,7 @@ import { ConnectedRouter } from 'connected-react-router'
 import { Provider } from 'react-redux';
 import { persistStore } from 'redux-persist';
 import { PersistGate } from 'redux-persist/integration/react';
+import { HelmetProvider } from 'react-helmet-async';
 
 import "./index.css";
 import App from "./App";
@@ -18,7 +19,9 @@ ReactDOM.render(
     <Provider store={store}>
       <ConnectedRouter history={history}>
         <PersistGate loading={null} persistor={persistor}>
-          <App />
+          <HelmetProvider>
+            <App />
+          </HelmetProvider>
         </PersistGate>
       </ConnectedRouter>
     </Provider>
